Continue solving warnings

+++ benches/bigint/bench_to_string.mojo
@@ -11,9 +11,9 @@ from decimo.tests import (
     print_header,
     print_summary_dual,
 )
-from python import Python, PythonObject
-from time import perf_counter_ns
-from collections import List
+from std.python import Python, PythonObject
+from std.time import perf_counter_ns
+from std.collections import List
 
 
 fn run_case(
@@ -41,21 +41,21 @@ fn run_case(
         var t0 = perf_counter_ns()
         for _ in range(iterations):
             _ = String(m1)
-        var t1 = (perf_counter_ns() - t0) / iterations
+        var t1 = (perf_counter_ns() - t0) / UInt(iterations)
         if t1 == 0:
             t1 = 1
 
         t0 = perf_counter_ns()
         for _ in range(iterations):
             _ = String(m2)
-        var t2 = (perf_counter_ns() - t0) / iterations
+        var t2 = (perf_counter_ns() - t0) / UInt(iterations)
         if t2 == 0:
             t2 = 1
 
         t0 = perf_counter_ns()
         for _ in range(iterations):
             _ = py.str(pa)
-        var tp = (perf_counter_ns() - t0) / iterations
+        var tp = (perf_counter_ns() - t0) / UInt(iterations)
 
         var s1 = Float64(tp) / Float64(t1)
         var s2 = Float64(tp) / Float64(t2)
